+++ src/Symfony/Component/HttpFoundation/RequestMatcher.php
@@ -29,6 +29,11 @@ class RequestMatcher implements RequestMatcherInterface
     private $host;
 
     /**
+     * @var int|null
+     */
+    private $port;
+
+    /**
      * @var string[]
      */
     private $methods = array();
@@ -56,13 +61,14 @@ class RequestMatcher implements RequestMatcherInterface
      * @param array                $attributes
      * @param string|string[]|null $schemes
      */
-    public function __construct(string $path = null, string $host = null, $methods = null, $ips = null, array $attributes = array(), $schemes = null)
+    public function __construct(string $path = null, string $host = null, $methods = null, $ips = null, array $attributes = array(), $schemes = null, int $port = null)
     {
         $this->matchPath($path);
         $this->matchHost($host);
         $this->matchMethod($methods);
         $this->matchIps($ips);
         $this->matchScheme($schemes);
+        $this->matchPort($port);
 
         foreach ($attributes as $k => $v) {
             $this->matchAttribute($k, $v);
@@ -87,6 +93,16 @@ class RequestMatcher implements RequestMatcherInterface
     public function matchHost($regexp)
     {
         $this->host = $regexp;
+    }
+
+    /**
+     * Adds a check for the the URL port.
+     *
+     * @param int|null $port The port number to connect to
+     */
+    public function matchPort(int $port = null)
+    {
+        $this->port = $port;
     }
 
     /**
@@ -164,6 +180,10 @@ class RequestMatcher implements RequestMatcherInterface
         }
 
         if (null !== $this->host && !preg_match('{'.$this->host.'}i', $request->getHost())) {
+            return false;
+        }
+
+        if (null !== $this->port && 0 < $this->port && $request->getPort() !== $this->port) {
             return false;
         }